requirements.txt
discord.py
python-dotenv
youtube_dl
requests
fastapi
uvicorn